chore(insights): extract date-range mapping utils, drop quill aliases
selectionForDateRange/dateRangeForSelection move to
lib/components/DateFilter/dateRangeSelection.ts with their tests; the
Quill* import prefixes were leftovers from when Lemon and quill imports
coexisted in the host file.
